requirements.txt
celery==5.4.0
django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-phonenumber-field==8.0.0
django-celery-beat==2.7.0
pillow==11.0.0
psycopg2-binary==2.9.10
python-telegram-bot==21.7
requests==2.32.3
uvicorn[standard]==0.32.0